+++ src/core/mod.rs
@@ -1,0 +1,7 @@
+/*!
+The `core` module is for types and concepts that are *not* strictly related to MIDI *files*.
+These types and concepts could be used for realtime MIDI as well.
+!*/
+
+mod clocks;
+mod duration_name;
